Add more restarttries for rutorrrent
Had a problem that rutorrent would exit with errorcode 255
multiple times and getting to the FATAL state
This would lead to effectively the container beeing useless.
I was able to start rutorrent again inside the container with
supervisorctl.
I hope this mitigates that.
Ideally would be of course a solution where supervisord exits
when a process enters the FATAL state.
A variable (or increasing) delay when restart might also help